Rollback changes to the origin
