--- app/src/main/java/androidlab1/database/Converters.kt
@@ -1,16 +1,0 @@
-package com.senex.androidlab1.database
-
-import androidx.room.TypeConverter
-import java.util.*
-
-class Converters {
-    @TypeConverter
-    fun fromTimestamp(value: Long?): Date? {
-        return value?.let { Date(it) }
-    }
-
-    @TypeConverter
-    fun dateToTimestamp(date: Date?): Long? {
-        return date?.time?.toLong()
-    }
-}
