+++ app/src/main/java/com/senex/androidlab1/views/fragments/SampleFragment.kt
@@ -1,0 +1,29 @@
+package com.senex.androidlab1.views.fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.senex.androidlab1.databinding.FragmentSampleBinding
+
+class SampleFragment : Fragment() {
+    private var _binding: FragmentSampleBinding? = null
+    private val binding
+        get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentSampleBinding.inflate(inflater, container, false)
+
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
