--- app/src/main/res/layout/list_item_user.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<com.google.android.material.card.MaterialCardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    app:cardCornerRadius="0dp"
-    app:cardElevation="2dp">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_margin="@dimen/size_double">
-
-        <com.google.android.material.imageview.ShapeableImageView
-            android:id="@+id/list_item_user_image"
-            android:layout_width="42dp"
-            android:layout_height="42dp"
-            android:scaleType="centerCrop"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:shapeAppearanceOverlay="@style/ShapeAppearanceOverlay.App.CornerSize50Percent"
-            tools:src="@drawable/image_profile_4" />
-
-        <TextView
-            android:id="@+id/list_item_user_nickname"
-            style="@style/text_view_style.header"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/size_double"
-            app:layout_constraintStart_toEndOf="@id/list_item_user_image"
-            app:layout_constraintTop_toTopOf="parent"
-            tools:text="Nickname" />
-
-        <ImageView
-            android:id="@+id/list_item_user_verification_icon"
-            android:layout_width="18dp"
-            android:layout_height="18dp"
-            android:layout_marginStart="@dimen/size_small"
-            android:layout_marginBottom="2dp"
-            android:src="@drawable/ic_check_24"
-            app:layout_constraintBottom_toTopOf="@+id/list_item_user_status"
-            app:layout_constraintStart_toEndOf="@+id/list_item_user_nickname"
-            app:layout_constraintTop_toTopOf="parent"
-            app:tint="@color/primary" />
-
-        <TextView
-            android:id="@+id/list_item_user_status"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="@dimen/size_double"
-            android:layout_marginEnd="@dimen/size_double"
-            app:layout_constrainedWidth="true"
-            app:layout_constraintEnd_toStartOf="@id/list_item_user_arrow_icon"
-            app:layout_constraintStart_toEndOf="@id/list_item_user_image"
-            app:layout_constraintTop_toBottomOf="@id/list_item_user_nickname"
-            tools:text="This message might be long" />
-
-        <ImageView
-            android:id="@+id/list_item_user_arrow_icon"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:src="@drawable/ic_arrow_right_24"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:tint="@color/primary" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</com.google.android.material.card.MaterialCardView>
