+++ app/src/main/res/navigation/nav_graph_main.xml
@@ -3,28 +3,11 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph_main"
-    app:startDestination="@id/listFragment">
+    app:startDestination="@id/sampleFragment">
 
     <fragment
-        android:id="@+id/listFragment"
-        android:name="com.senex.androidlab1.views.fragments.ListFragment"
+        android:id="@+id/sampleFragment"
+        android:name="com.senex.androidlab1.views.fragments.SampleFragment"
         android:label="SampleFragment"
-        tools:layout="@layout/fragment_list">
-        <action
-            android:id="@+id/action_listFragment_to_infoFragment"
-            app:destination="@id/infoFragment" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/infoFragment"
-        android:name="com.senex.androidlab1.views.fragments.InfoFragment"
-        android:label="InfoFragment"
-        tools:layout="@layout/fragment_info">
-        <argument
-            android:name="userId"
-            app:argType="long" />
-        <action
-            android:id="@+id/action_infoFragment_to_listFragment"
-            app:destination="@id/listFragment" />
-    </fragment>
+        tools:layout="@layout/fragment_sample"/>
 </navigation>
